Abstract existing submit order form into separate views with individual CTAs
